requirements.txt
flask
pytest
requests
pyspellchecker